added link for FAQ to editor
