+++ apps/client/components/Faq.tsx
@@ -2,6 +2,7 @@ import { Box, Paper, Stack, Typography } from "@mui/material";
 import * as React from "react";
 import FaqQuestion from "./buildingBlocks/FaqQuestion";
 import NavigationBar from "./NavigationBar";
+import { questions } from "../data/questions";
 
 type FaqProps = {
   parameterObject: {
@@ -18,28 +19,6 @@ type FaqProps = {
   };
 };
 export const Faq = ({ parameterObject }: FaqProps) => {
-  const questions: { question: string; answer: string }[] = [
-    {
-      question: "What is Sjorcraft?",
-      answer:
-        "Sjorcraft is a web-based 3D editor for creating, sharing and exporting SJORconstructions.",
-    },
-    {
-      question: "Can I export my construction from Sjorcraft?",
-      answer:
-        "Yes, you can export your construction from Sjorcraft as different files (e.g. .sjor, .jpg, .stl, .dae) different filetypes are on the way. You can find the export button in the top right corner of the editor.",
-    },
-    {
-      question: "Is Sjorcraft free to use?",
-      answer:
-        'Yes, Sjorcraft is free to use. You can always support us by "buying us a coffee" on our website.',
-    },
-    {
-      question: "Can I cancel drawing a pole?",
-      answer:
-        "Yes, you can cancel drawing a pole by rightclicking while drawing. This action is possible while using the pole-, bipod- and tripodtool.",
-    },
-  ];
   return (
     <Box>
       <NavigationBar parameterObject={parameterObject} />
@@ -56,7 +35,7 @@ export const Faq = ({ parameterObject }: FaqProps) => {
         <Typography
           variant="h1"
           sx={{
-            fontSize: { xs: "2rem", sm: "3rem", md: "4rem" },
+            fontSize: { xs: "2rem", sm: "2.5rem", md: "3rem" },
             mb: 2,
           }}
           color={"primary.contrastText"}
